# backend/src/routes/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"RegisterUserRequest": {
			"properties": {
				"name": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"contact_name": {
					"type": "string"
				},
				"address": {
					"type": "string"
				},
				"postal_code": {
					"type": "string"
				},
				"city": {
					"type": "string"
				},
				"user_type": {
					"type": "string"
				},
				"state": {
					"type": "string"
				},
				"country": {
					"type": "string"
				},
				"company_name": {
					"type": "string"
				},
				"gst_number": {
					"type": "string"
				},
				"phone": {
					"type": "string"
				},
				"subscription_type": {
					"type": "string"
				}
			},
			"required": [
				"name",
				"email",
				"password",
				"contact_name",
				"address",
				"postal_code",
				"city",
				"user_type",
				"state",
				"country",
				"company_name",
				"gst_number",
				"phone",
				"subscription_type"
			],
			"type": "object",
			"additionalProperties": false
		},
		"LoginUserRequest": {
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"password"
			],
			"type": "object",
			"additionalProperties": false
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"statusCode": {
									"type": "number",
									"format": "double"
								},
								"message": {
									"type": "string"
								}
							},
							"required": [
								"message"
							],
							"type": "object"
						}
					}
				},
				"consumes": [
					"application/json"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RegisterUserRequest"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"properties": {
								"statusCode": {
									"type": "number",
									"format": "double"
								},
								"token": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}
							},
							"required": [
								"message"
							],
							"type": "object"
						}
					}
				},
				"consumes": [
					"application/json"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LoginUserRequest"
						}
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}